+++ src/beingequipbackend.cpp
@@ -39,7 +39,7 @@ BeingEquipBackend::BeingEquipBackend(Being *const being)
         for (size_t f = 0; f < sz; f ++)
         {
             const int idx = Net::getInventoryHandler()->
-                convertFromServerSlot(f);
+                convertFromServerSlot(static_cast<int>(f));
             const int id = being->mSpriteIDs[f];
             if (id > 0 && idx >= 0 && idx < EQUIPMENT_SIZE)
             {
